+++ src/OpenClawAgent.Service/Program.cs
@@ -53,7 +53,7 @@ try
     builder.Services.AddHostedService<RemediationPoller>();
     builder.Services.AddHostedService<AutoUpdater>();
     builder.Services.AddHostedService<LiveDataPoller>();
-    builder.Services.AddHostedService<Screen.ScreenStreamService>();
+    builder.Services.AddHostedService<OpenClawAgent.Service.Screen.ScreenStreamService>();
 
     var host = builder.Build();
     host.Run();